requirements.txt
asgiref==3.8.1
django==5.2
sqlparse==0.5.3
tzdata==2025.2
requests==2.32.3
djangorestframework==3.16.0
